requirements.txt
google-generativeai
fastapi